move usage to common, rename to accomodate
