+++ common-lib/src/main/scala/com/gu/mediaservice/model/UsageSource.scala
@@ -1,0 +1,13 @@
+package com.gu.mediaservice.model
+
+import play.api.libs.json._
+
+
+case class UsageSource(
+  usageType: String,
+  uri: Option[String] = None,
+  name: Option[String] = None
+)
+object UsageSource {
+  implicit val writes: Writes[UsageSource] = Json.writes[UsageSource]
+}
